TA Facial Recognition
Got DeepFace to work with webcam in login screen. I used flask to host a server so the express backend can take a submitted picture and picture from the database and send it to the flask server, which would run DeepFace verification on them.

Still Need to make it so it logs TAs into the clockIn page. To do so I will most likely need to create a new clockIn page that does not require supabase authentication. I have started on that with FaceClockIn.tsx.